put all Plugin-Configuration into its own files
