+++ budget-binder-server/src/main/kotlin/de/hsfl/budgetBinder/server/plugins/Logging.kt
@@ -1,0 +1,12 @@
+package de.hsfl.budgetBinder.server.plugins
+
+import io.ktor.server.application.*
+import io.ktor.server.plugins.callloging.*
+import org.slf4j.event.Level
+
+fun Application.configureLogging() {
+    install(CallLogging) {
+        level = Level.INFO
+        disableDefaultColors()
+    }
+}
